glb-redirect: support path MTU discovery
Path MTU discovery requires that Packet Too Big messages are delivered
to the machine "responsible" for a given flow. Modify glb-redirect
so that it reconstructs the original flow tuple from the ICMP message,
and make it pass such messages along the chain.

Add a helper function which creates a fake TCP header from
the ICMP payload, adjusting for the direction of the flow.
This is then used to check whether a connection is known locally.

Fixes #47